Moved a few model classes from logic/ into logic/model/.
