+++ src/logic/models/LibraryModel.h
@@ -38,12 +38,12 @@ class QFileDevice;
 #include <logic/serialization/ISerializable.h>
 #include <logic/dbmodels/CollectionDatabaseModel.h>
 #include <concurrency/ThreadsafeMap.h>
-#include <models/ColumnSpec.h>
-#include "Library.h"
-#include "LibraryRescanner.h" ///< For MetadataReturnVal
-#include "LibraryEntry.h"
-#include <jobs/LibraryEntryLoaderJob.h>
-#include "LibraryRescannerMapItem.h"
+#include <ColumnSpec.h>
+#include "logic/Library.h"
+#include "logic/LibraryRescanner.h" ///< For MetadataReturnVal
+#include "logic/LibraryEntry.h"
+#include <logic/jobs/LibraryEntryLoaderJob.h>
+#include "logic/LibraryRescannerMapItem.h"
 #include <utils/RegisterQtMetatypes.h> ///< For common metatype declarations of C++ std types.
