+++ src/logic/models/PlaylistModel.cpp
