+++ src/logic/models/PlaylistModelItem.cpp
