+++ src/logic/proxymodels/PlaylistSortFilterProxyModel.cpp
