+++ src/logic/proxymodels/PlaylistSortFilterProxyModel.h
